Add notifications and move email logic
All of the legwork involving email is happening behind the scenes so it
makes sense to move those methods out of the controller.  With user
notifications the issue was keeping them on a rolling schedule on a per
user basis, which means we need access to the timestamp on the last
notification associated with a particular user.